Reorganize wasm-rust directory to be more cargo-friendly.
