+++ src/transactions/test/wasm-rust/Makefile
@@ -1,13 +1,6 @@
-.PHONY: build
-
-build: test_invoke_contract.wasm test_typed_contract.wasm test_other_contract.wasm
-
-
-STRIP=symbols
-#STRIP=debuginfo
-
-%.wasm: %.rs Makefile $(wildcard src/*.rs)
-	rustc --edition 2021 --crate-type cdylib --target wasm32-unknown-unknown -C opt-level=z -C panic=abort -C overflow-checks=no -C strip=$(STRIP)  $<
+all:
+	cargo build --target wasm32-unknown-unknown --release
 
 clean:
-	rm -f *.wasm
+	cargo clean
+
