+++ src/transactions/test/wasm-rust/test_call_helper/src/lib.rs
@@ -1,0 +1,15 @@
+#![no_std]
+use stellar_contract_sdk as sdk;
+use sdk::{Val};
+
+mod helper_contract_interface;
+use helper_contract_interface::HelperContract;
+
+// Here we're using the typesafe HelperContract interface
+// to compose a call to another contract (whose LedgerKey
+// identity is passed to us as an argument)
+
+#[no_mangle]
+pub fn call(contract: HelperContract) -> Val {
+    contract.helper(10)
+}
